requirements.txt
Django==1.11
djangorestframework==3.10.2
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.2
